pyproject.toml
[tool.poetry]
name = "conf-manager"
version = "0.1.0"
description = "A tool to manage configuration file overrides"
authors = ["Johnny Castrup <johnny@castrup.net>"]
readme = "README.md"

packages = [{include = "conf_manager"}]

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
conf-manager = "conf_manager.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "conf_manager"]